test: split settings renderers
